createUser moved to user controller
